requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
Django==5.2.12
et_xmlfile==2.0.0
exceptiongroup==1.3.1
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
h11==0.16.0
httplib2==0.31.2
idna==3.11
numpy==2.2.6
oauthlib==3.3.1
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.3.3
proto-plus==1.27.1
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.32.5
requests-oauthlib==2.0.0
selenium==4.41.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.5
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
